+++ Changes.md
@@ -9,7 +9,7 @@ Example:
     "unsuppress": []
   },
 ```
-
+- Fix issue with punned record expressions in OCaml projects.
 
 # 2.19.0
 - Don't report redundant optional arguments on functions annotated `@live`.